No longer store generated documentation markdown in this repo
The amount of files we generated on a per package basis is really high.
As we move forward with having a large number of packages hosted on
main.actor, it will become hard to work efficiently with the code in this
repo as it grows larger and larger with generated documentation.

This commit moves the generated markdown documentation from being added
to and stored in this repo and instead moves it to being stored in
https://github.com/ponylang/main.actor-package-markdown.

Closes #18