local gatsby install
Because we have gatsby as a local dep in package.json, we don't need
to install it globally (in fact the global install just proxies to the
local anyway, so we're doing double work). the package.json has a
`build` script, which has access to all of the binaries in `node_modules/bin` including `gatsby`. We can use npm to trigger this build script.